requirements.txt
# Web Scraping
beautifulsoup4==4.13.3
scrapy==2.11.0
selenium==4.27.1
webdriver-manager==4.0.2
googlemaps==4.10.0
# Core Data Libraries
pandas==1.5.3
numpy==1.23.5
scipy==1.10.1
python-dateutil==2.8.2
pyyaml==6.0.1
tqdm==4.67.1
requests==2.31.0
python-dotenv==1.0.1

# NLP & Text Processing
nltk==3.9.1
spacy==3.7.2
gensim==4.3.3
langdetect==1.0.9
textblob==0.17.1
scikit-learn==1.3.2
torch==2.4.1
tokenizers==0.20.3
huggingface-hub==0.29.3
transformers==4.46.3

# Database
psycopg2-binary==2.9.10
sqlalchemy==1.4.51

# DBT (Data Build Tool)
dbt-core==1.7.12
dbt-postgres==1.7.12

# Airflow & Scheduling
apache-airflow==2.8.1
apache-airflow-providers-postgres==5.8.0
apache-airflow-providers-http==5.10.0
